deqp: Rerun if tests fail to connect to X
A pretty common error it seems is to get a failure to open display when
using X11. This adds an extra method to the DEQPBaseTest class that
reruns the test is it finds that error in the stderr. This affects only
deqp based suite integration.

This fixes a number of sporadically failing tests from our CI.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>
Reviewed-by: Mark Janes <mark.a.janes@intel.com>