arb_program_interface_query: test stage ref for buffer blocks
This tests that we get the correct stage reference information when
a shader has both a shader storage and uniform block.

This exposes a bug in Mesa where we were expecting the index to be
an index of a combined SSBO and UBO resources list. When they should
instead each have their own resource list and indices.