arb_vertex_attrib_64bit: add test zero-vertex-attrib
ARB_vertex_attrib_64bit spec specifies that calling GetVertexAttribLdv
with pname CURRENT_VERTEX_ATTRIB and index 0 should raise a
INVALID_OPERATION error.

That is also true for the rest of GetVertexAttrib* methods depending
on the spec you read, as some doesn't mention it explicitly, other do
in core and compatibility and other just on compatibility. Current mesa
code would only raise the error on compatibility mode, that would
make this test somewhat useless.

FIXME:
Asked on piglit list about to ask spec clarification:
https://lists.freedesktop.org/archives/piglit/2016-April/019348.html

But it is possible that it would be better to just not add this test.