tests/all.py: Use TESTS_DIR for functions that don't have it
TESTS_DIR is an absolute path that points at <piglit directory>/tests,
which is needed to make piglit work when built out of tree (ie, when
installed).

This fixes a number of tests That assume piglit will be run in the root
dir, to add TEST_DIR to their os.path.join calls. This makes them work
when run from an installed piglit.

cc: thomas.stellard@amd.com
Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>