unittests: Fix tests that leave temporary files laying around
These are a bit more complicated, and the solution is more difficult.
Basically the problem is that the temporary file gets renamed out from
under the test, and the name of the file is a generated and unknown. To
get rid of these I've wrapped the shared private method in the
utils.test_in_tempdir, and extended that decorator to return a value.

This is kind of a hack, and all of this should be fixed by mocking out
things rather than using files.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>