generators: Add a simple generator for enabled and disabled defines
This replaces a large number of handrolled tests (and adds many new
ones) for testing that preprocessor defines are exposed correctly
(either exposed if the extension is enabled, or not if it's disabled)
across all shader stages. It tests for support in legacy (non
core/compat) mode and in profile mode, unless the extension itself
requires GLSL >= 140. It also covers all stages (fs, vs, gs, tcs, tes,
and cs), and both GLES and GL.

It drives all of this from a simple, easily extended list in the
generator, and replaces over 1000 lines of code with about 300,
including the template files, and generates over 1500 tests currently.

These are GLSL parser tests and don't take more than a few seconds to
run.

The only known issue is that an extension can not currently be tested in
both GLES and OpenGL.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>
Acked-by: Ilia Mirkin <imirkin@alum.mit.edu>