python: Make mako template dirs robust
This puts the templates in a directory structure that makes them robust
to both changes in python version and mako version.

This fixes errors caused by downgrading mako, which isn't that common,
unless you're running the unit tests with use the oldest supported mako,
and may conflict with the compiled templates from the system mako.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>