arb_texture_multisample-stencil-clear: new stencil clear test
This exhibits an MSAA-specific bug in the radeonsi driver with fast stencil
clear and decompress.

The test can optionally disable MSAA and change the number of samples, though
I've only added the default variant to the all run - that's enough to expose
the bug.

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=93767