glx_arb_sync_control/wait_for_msc: Add a missing timeout
This test requests via GLX to wait until a certain vertical retrace
event (based on the Media Stream Counter, AKA MSC) before being
unblocked by the X-Server. Since there can be bugs in the
implementation, it is possible for the test to not terminate.
Re-enforcing this idea is that all the other glx_arb_sync_control
tests already have a timeout set.

This change adds a timeout to guard against this case and report a failure.

V2: Only added timeout to waitformsc and updated commit message.

Signed-off-by: Marta Lofstedt <marta.lofstedt@linux.intel.com>
Reviewed-by: Martin Peres <martin.peres@linux.intel.com>