glx-query-drawable: Relax a check about bare windows
The map from visuals to fbconfigs is injective; there may be multiple
fbconfigs that are compatible with a particular visual. GLXWindows are
created with respect to a (single) fbconfig, but Windows are not, and
the implementation may reasonably not have an answer in the latter case
to the question "what fbconfig is associated with this window".

This effectively blesses the behavior of Xorg and NVIDIA in this case.

Signed-off-by: Adam Jackson <ajax@redhat.com>