Test that glShaderSource does not change compile status.
OpenGL 4.5 Core Profile section 7.1, in the documentation for
CompileShader, says: "Changing the source code of a shader object with
ShaderSource does not change its compile status or the compiled shader
code."

This test creates a shader, compiles it, changes its source, and links
it. The spec requires rendering done with this shader to be consistent
with the old source, not the new source, since the shader isn't compiled
again after the source is changed.

Signed-off-by: Jamey Sharp <jamey@minilop.net>
Reviewed-by: Kenneth Graunke <kenneth@whitecape.org>