framework/profile: Convert TestDict to a collections.MutableMapping
MutableMapping is an abstract base class that makes writing dict-like
objects easier, by providing some of the boilerplate code. This is the
recommended way to make a dict-like object, as opposed to subclassing
dict. This is because subclassing built-ins is perilous, sometimes they
call their protocol methods (or magic methods), sometimes they don't.

This change uncovered such a bug, TestDict.update would silently
overwrite other tests when calling update, without there being an code
to explicitly do that. This patch additionally fixes the test that trips
the bug and adds a test the explicitly cover it.

This is groundwork for the next patch.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>
Acked-by: Brian Paul <brianp@vmware.com>
Tested-by: Brian Paul <brianp@vmware.com>
Tested-by: Gabriel Feceoru <gabriel.feceoru@intel.com>