arb_internalformat_query2: use PRIi64 instead of PRIu64 macro
As all the values we are printing are 64 bit integers, not unsigned integers.

In general this is not relevant, as no query for a supported combination
of pname,target and internalformat will return a negative value. But for
pnames like SAMPLES, if not supported, params should remain unmodified. So
if the buffer were initialized to a negative value for some reason, the
value would be wrongly printed.

Reviewed-by: Eduardo Lima Mitev <elima@igalia.com>