builtin_function*py: more test data for "trunc"
When performing a trunc op in a float value there is, basically, three
scenarios: when the exponent is < 0, when it is > bitsize of the
significand or with a value in the middle. Current generator was using
data for values in the middle and < 0 but not for the last case so we
add now a couple of values in that range and also the interesting 0.0
special value.

Reviewed-by: Eric Anholt <eric@anholt.net>
Reviewed-by: Dylan Baker <dylan@pnwbakers.com>
Signed-off-by: Andres Gomez <agomez@igalia.com>