framework/profile: don't replace TestDict instance with regular dict
Currently when the test dictionary is filtered the TestDict instance is
replaced by a built-in-dict instance. This isn't ideal for a number of
reasons, obviously the silent change in type isn't good. This is also
groundwork for other patches in this series.

This also fixes a number of unit tests that I'm not exactly sure why
they passed before this test. They were doing some invalid things,
mainly mangling types. They also happened to be cloning data after
destructively editing the data. They are fixed now.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>
Acked-by: Brian Paul <brianp@vmware.com>
Tested-by: Brian Paul <brianp@vmware.com>
Tested-by: Gabriel Feceoru <gabriel.feceoru@intel.com>