framework/profile: Make test run order deterministic
Test run order has been non-deterministic because of the data structure
used, a dict. This patch simple replaces the dict with an OrderedDict,
which is just like a dict, except that it remembers insertion order.
This means that when using -1/--no-concurrency option that tests will be
run in a deterministic order.

What this does not change, however, is the sorting of the tests in the
on disk JSON file, those remain unsorted.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>
Acked-by: Brian Paul <brianp@vmware.com>
Tested-by: Brian Paul <brianp@vmware.com>
Tested-by: Gabriel Feceoru <gabriel.feceoru@intel.com>