framework: Make JSON backend test ordering deterministic
This changes the JSON backend to write results in a deterministic order
in the JSON store (assuming that concurrency is not used, there's no
easy way to enforce this with concurrency), this will be the same order
that the tests are run in. In combination with the previous patch this
will result in tests being run and written in a set order.

It's worth noting that because the JUnit backend uses different data
structures (list like objects) instead of dict-like objects it has
always had this property.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>
Acked-by: Brian Paul <brianp@vmware.com>
Tested-by: Brian Paul <brianp@vmware.com>
Tested-by: Gabriel Feceoru <gabriel.feceoru@intel.com>