framework: Make testlist option impose ordering
The --test-list argument hasn't been used a lot because it isn't that
useful. Currently it is just a list of values passed into the same
function that handles the -t/--include-tests switch. That isn't very
very useful since -t can be given multiple times, and it is regex so
it's easy enough to build some complex regular expressions.

This patch changes the behavior of test to an authoritative list of
tests that will be run in a defined order (or as close as possible when
using concurrency). It does not exclude the -t and -x options, and can
be run with them, retaining ordering as much as possible. This means
that passing a testlist will remove any tests from the profile that are
not in the list (it also enforces that all of the tests exist), and runs
them in the order of the testlist file.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>
Acked-by: Brian Paul <brianp@vmware.com>
Tested-by: Brian Paul <brianp@vmware.com>
Tested-by: Gabriel Feceoru <gabriel.feceoru@intel.com>