builtin_function*py: Corrected single to double conversions
Calculation of the tolerance in single precision with very high
numbers was giving "inf" results while it was possible to calculate in
double precision and round back to single precision afterwards. In
addition, since the vector values are already calculated in single
precision, there is no need for additional conversions.

For the double precision module, we don't really need to do any
conversion since the calculations are already done in double
precision. Additionally, we removed custom functions that are not used
any more.

Signed-off-by: Andres Gomez <agomez@igalia.com>
Reviewed-by: Eduardo Lima Mitev <elima@igalia.com>