framework: Don't report an error message for fast skipping
Piglit has a number of mechanisms for fast skipping, some for GL
features, some for platform support, and some other things. It puts a
nice little message in stdout to let the developer know why it skipped,
but currently that message includes something to the effect of "This is
an unhandled exception", which it isn't it's being handled
appropriately. This patch fixes that.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>