arb_internalformat_query2: print properly the value of a failing case
Before this commit, it was printed the numerical value and always tried
to get a GL enum from it. This commit checks if the pname returns a
enum or not and decide.

Additionally it also do some wrapping on the fact that GL_NONE, GL_POINTS
and GL_FALSE has the same value.

Acked-by: Eduardo Lima <elima@igalia.com>
Acked-by: Antia Puentes <apuentes@igalia.com>