framework/backends/junit: Fix invalid JUnit output
In commit b0d05323e code was added to produce clearer error messages for
tests who's status changed from crash to fail and vice versa. There is a
not so subtle bug in that patch, it adds a "crash" element, but that
element should be an "error" element.

This patch fixes that bug.

cc: Kenneth Graunke <kenneth@whitecape.org>
Reviewed-by: Mark Janes <mark.a.janes@intel.com>
Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com