shaders: port "GLSL link single global initializer, 2 shaders" to shader_runner
glsl-link-test is a very simple program, but all of it's tests can be
implemented as shader_runner tests. It also has an annoying habit of
not working when running out of tree. Rather than trying to fix that bug
just start porting its few tests to shader_runner to delete it.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>
Reviewed-by: Kenneth Graunke <kenneth@whitecape.org>