framework/results: Fix subtests masking crashes
In piglit when a test generates subtests we treat the "worst" subtest as
the test status, and in most cases this works as expected. There is at
least one case where this is not correct, and that is the case of crash.

There are a couple of reasons that crash should not be masked. One is
that it is generated by the framework when the test binary hits an
assert or segfaults, or any number of similar cases. The second is that
it may mean that all of the subtests did not run, as such we don't want
the status to be masked by the "worst" subtest which would be fail at
worst.

This is maybe not the strictly correct way to handle this. Probably the
better solution is to include the overall result in the max call if the
overall result is not NOTRUN. But this fixes the problem and that can be
done later.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>
Reviewed-by: Brian Paul <brianp@vmware.com>
Tested-by: Brian Paul <brianp@vmware.com>