arb_internalformat_query2: being more strict with the format/type inferred from internalformat
For max-dimensions test we are trying to create a texture based on the
target and internalformat, in order to confirm if it is supported or not.

gles seems to be more strict on the type/format used, so we need to
be more careful.

This commit helps to get max-dimensions test passing on gles.

Acked-by: Eduardo Lima <elima@igalia.com>
Acked-by: Antia Puentes <apuentes@igalia.com>