arb_internalformat_query2: 0 is a valid value for NUM_SAMPLE_COUNTS
minmax is testing that the implementation fulfill the minimum values
stablished by the spec. Copied from the equivalent arb_internalformat_query1
test, it was returning error if any of the tested formats returned 0 for
NUM_SAMPLE_COUNTS, as the minimum value was 1. That is not a restriction
for arb_internalformat_query2 anymore, where 0 is a valid value. Error on
negative values are still being tested.

Acked-by: Eduardo Lima <elima@igalia.com>
Acked-by: Antia Puentes <apuentes@igalia.com>