arb_internalformat_query2: add common config macro for all query2 tests
As we are using the same source code in order to test this extension
on desktop gl, gles2 or gles3, the values of supports_gl_es_version or
supports_gl_compat_version can vary. Instead of using #ifdefs, one
alternative would be pass this value as a command line parameter
when calling the binary.

Acked-by: Eduardo Lima <elima@igalia.com>
Acked-by: Antia Puentes <apuentes@igalia.com>