Add more ARB_cull_distance tests
This adds a handful simpler tests for ARB_cull_distance. Instead of the
uncessarily complicated math in the old tests, we just set the distances
based on gl_VertexID. That way we can reliably select which triangles to
cull and/or clip and where we clip them.

We also replace fs-cull-and-clip-distance-exceed-max.shader_test with a
C test so we can properly catch the specified compile or link error when
one of the array sizes or the sum of their sizes exceed the limits.

Reviewed-by: Dave Airlie <airlied@redhat.com>