unittests: fix dmesg test that could not work on non-posix platforms
The test would try to initialize a linuxDmesg even on windows, which
obviously wouldn't work, since there is not /dev/kmesg on windows. The
test is a little hacky anyways, and better tested using appveyor for
windows testing. The updated test just tests that the currently platform
returns the right thing.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>