unittests: reimplement utility function to get a unique file to be windows-safe
tempfile.NamedTemproraryFile doesn't work in all of the cases that it
needs to for piglit's unittests when running on windows (though it does
on POSIX systems), so this reimplements it to use a unique temporary
directory with a known filename. This serves the same purpose, but
doesn't break on windows.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>