shader_runner: Fix get_ints on 32-bit systems.
The new ARB_vertex_attrib_64bit tests specify integer uniform values
as hex, such as 0xc21620c5.  As an integer value, this is beyond LONG_MAX
on 32-bit systems.  The intent is to parse it as an unsigned hex value and
bitcast it.

However, we still need to handle parsing values with negative signs.

Using strtoll and truncating works.

v2: Drop hunks that should have been in patch 1.

Signed-off-by: Kenneth Graunke <kenneth@whitecape.org>
Reviewed-by: Antia Puentes <apuentes@igalia.com>
Reviewed-by: Andres Gomez <agomez@igalia.com>
Tested-by: Mark Janes <mark.a.janes@intel.com>