generators/gen_vs_in_fp64: Rewrite the columns test generator
This rewrites the columns generator to use a single template rather than
the inherited template. The net result is that these tests are generated
in 1/3 of the time.

These are the minority tests from this generator. I measure that they
take about .3 seconds to generate (with this patch). So while this is
saving us a little bit of time, it's not going to make a big difference.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>
Reviewed-by: Andres Gomez <agomez@igalia.com>