generators/gen_vs_in_fp64: reimplement require_execution template
This cuts the runtime for this generator by 3/4 on my system. That's 4
minutes to 1 minute. That's pretty impressive, and for both my system
and for our CI system this is a nice improvement. The generated files
are the same, minus whitespace changes (diff -NaurBw notes no
differences).

While I think the new template is more readable than the original code,
and it is definitely faster, it does introduce more code duplication
that could be cleaned up in a follow on patch.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>
Reviewed-by: Andres Gomez <agomez@igalia.com>