generators/modules: Add GLSL types representations
This adds a new module with glsl type representations. The immediate
plan it to use these in the gen_vs_in_fp64 generator to reduce runtime
by removing the need for the cols and rows functions, which are
expensive.

These may be applicable to other generators, but those will have to
wait.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>
Reviewed-by: Andres Gomez <agomez@igalia.com>