egl: new test egl-create-msaa-pbuffer-surface
Test initializes EGLConfig with EGL_SAMPLES set to 4 and attempts
to create a pbuffer with such config. This causes a crash in Mesa
i965 driver but passes on Nvidia binary driver version 355.11.

v2: set pbuffer surface type to test attribs (for eglChooseConfig)
    skip if test fails already when choosing config as such config
    may not be supported

(dEQP EGL tests are currently hitting this path)