util: Wrapper to load exact hex patterns for ints
For some cases we want to have shaders where we load an exact bit
pattern into a signed int.

If we would just use strtol, the negative values would have to include
the sign, as in -0x7fffffff, and not the exact bit pattern, which
would be 0x80000000, to avoid an ERANGE out of range errno value on
32-bit systems.

Hence, we were using strtoll but that was also causing that we would
lose the internal range check performed by strtol.

Now, we use an implementation that would accept exact bit patterns,
including a proper internal range check and, also, the other
alternative formats allowed by strtol.

Signed-off-by: Andres Gomez <agomez@igalia.com>
Reviewed-by: Alejandro Piñeiro <apinheiro@igalia.com>