util: specify channel sizes for choose_config_attribs()
WAFFLE_RED_SIZE=1 (for example) gets passed to glXChooseFBConfig as
GLX_RED_SIZE=1.  GLX chooses the deepest pixel format which meets that
minimum value so we typically get an 8 bit/channel format.

But with wglChoosePixelFormatARB() we may get the shallowest pixel
format that meets that value so we may wind up getting a 5/6/5 format.
Many piglit tests aren't prepared for that and report failures because
the color tolerances are too tight.  In any case, we really want to test
8-bit formats.

This change tells waffle to first try 8/8/8 color, 24-bit Z.  And if
that fails, try 5/6/5 with 16-bit Z.

Reviewed-by: Jose Fonseca <jfonseca@vmware.com>