GLX_OML_sync_control: remove '#if defined(GLX_MESA_swap_control)' guards
Analogous to previous commit - we don't want the test to differ between
builds. This could lead to the test silently skipping when built against
old headers, _regardless_ if at runtime the extension is supported or
not.

With this commit no GLX test in piglit (barring glean) relies on this
broken behaviour. Checked with the following:

$ git grep "defined.*\<GLX_"
$ git grep "ifdef.*\<GLX_"

Only local definitions can be seen with the inverse query
$ git grep "ifndef.*\<GLX_"

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>
Reviewed-by: Ilia Mirkin <imirkin@alum.mit.edu>