Generators: gen_shader_image_load_store_tests: remove duplicate tests
This generator produces preprocessor defines tests, which are duplicated
by the gen_extensions_defined generator. It makes sense to me to let the
defines generator own this functionality (it's very minor anyway).

This also changes the version required in extensions defined to 130, per
the spec "OpenGL 3.0 and GLSL 1.30 are required."

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>