Generators/genclbuiltins: Use open as context manager to ensure files are closed
Currently there are a good number of files that aren't closed because
the call to the close function is in the wrong loop level. While moving
it in one level would solve the problem, using a context manager is
better style and ensures that the file is closed.

This error was spotted by py.test.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>