util: Add u/byte, u/short and half support to [vertex data]
This allows to set data of u/byte, u/short and half types for
attributes with the shader runner.

For example:
0/byte/int attname1/ushort/uint attname3/half/float

The syntax has been extended so the recommended way has replaced the
old COUNT field in the [vertex data] header line with the
corresponding GLSL type for the old TYPE field.

In any case, the extended syntax is backward compatible so it is still
possible to specify a vec3 attribute like:
attname/float/3

In addition to the now recommended format:
attname/float/vec3

Due to this, the arb_vertex_attrib_64bit input tests generator has
been also adapted to the new syntax.

Signed-off-by: Andres Gomez <agomez@igalia.com>
Reviewed-by: Alejandro Piñeiro <apinheiro@igalia.com>