arb_copy_image-api_errors: fix one error check (INVALID_ENUM instead of INVALID_OPERATION)
This patch adapts the test to the last change on mesa. Copying
the rationale on that patch here:

For the case (both src or dst) where we had a texobject, but the
texobject target was not the same that the method target, this spec
paragraph was appplied:

 /* Section 18.3.2 (Copying Between Images) of the OpenGL 4.5 Core
  * Profile spec says:
  *
  *     "An INVALID_VALUE error is generated if either name does not
  *     correspond to a valid renderbuffer or texture object according
  *     to the corresponding target parameter."
  */

But for that case, the correct spec paragraph should be:
 /* Section 18.3.2 (Copying Between Images) of the OpenGL 4.5 Core
  * Profile spec says:
  *
  *     "An INVALID_ENUM error is generated if either target is
  *      not RENDERBUFFER or a valid non-proxy texture target;
  *      is TEXTURE_BUFFER or one of the cubemap face selectors
  *      described in table 8.18; or if the target does not
  *      match the type of the object."
  */

specifically the last sentence: "or if the target does not match the
type of the object".

This patch fixes the error checked (s/INVALID/ENUM) for that case,
and moves down the INVALID_VALUE spec paragraph, as that case (invalid
texture object) was already checked on the test, but without spec
paragraph.

Reviewed-by: Nicolai Hähnle <nicolai.haehnle@amd.com>