framework: fix two bugs in status
The Status.__bytes__ method is broken in python3 since it doesn't
provide an encoding currently, and __repr__ should be returning a value
that can be used by exec() to recreate the class. To fix bytes we
actually need to have different paths for python 2 and python 3, since
python 2's str() doesn't have an encoding argument.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>