framework: Read test files as utf-8
Make sure test files are always considered utf-8 to avoid a
UnicodeDecodeError when e.g. locale is unset.

This requires using io.open() as python2 open() does not support
"encoding". The upside of this is that same code paths should now work
for both python2 and python3.

v2: Fix unittests that now need to mock io.open instead of
    __builtins__.open (Dylan)

Signed-off-by: Jussi Kukkonen <jussi.kukkonen@intel.com>
Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>
Reviewed-by: Dylan Baker <dylan@pnwbakers.com>