shader_runner: Drop unnecessary texture target argument from fb tex layered command.
The texture target argument of the 'fb tex layered' command is
somewhat redundant because the glFramebufferTexture() GL API it's
implemented in terms of doesn't care what the target is.  Right now
the implementation of 'fb tex layered' requires the texture to be a 2D
array due to the way it uses the GL API to query the bound texture
object and dimensions, but that limitation will be removed shortly.