framework: Add required option to deqp.get_option
Some options in deqp are essential, but we don't enforce that. Which
leads to the deqp profile being unable to get it's binary (for example)
and failing through until the framework excepts importing the suite.

This patch lays the ground-work for fixing that by adding a required
keyword argument, which when set to True causes a FatalError to be
raised if that option falls through to the default.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>