EGL_EXT_image_dma_buf_import/refcount: New test for GEM handle bug.
Drivers can miss the need to refcount their GEM handles, and the
destroy of the image/texture pair for tex2 will result in tex1 no
longer being usable.

This fails to piglit_report_result(PIGLIT_FAIL) on vc4 (which is
missing the handle code), because tex1's shadow tiled temporary
already has the texture contents in it when the shadow linear->tiled
blit fails, but the driver does make a noisy warning on stderr.  On
intel, if you drop the global name lookup in libdrm, the driver
exit()s due to the -EINVAL from the draw call.

v2: add to all.py.

Reviewed-by: Rob Clark <robdclark@gmail.com>