unittests: replace json_tests with a schema and validator
This module is really a want for a schema and enforcement for that
schema, since there is in fact schema for JSON and several python
libraries to enforce said schema, it would be much better to use use one
of those, since that also makes the output format reproducible and
portable.

This adds the schema and the proper test, which currently this test
xfails because bugs.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>