framework/test/glsl_parser_test: Add ARB_ESx to command line when needed
When using the GL binary to run GLES tests the appropriate ES
compatibility extension is added to the fast skipping requirements, but
not to the command line options. It needs to be added to the command
line options too, so that it will be skipped properly when not using the
fast-skipping layer.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>