Change arb_framebuffer_srgb/blit to follow GL 4.4+ rules.
OpenGL originally never did sRGB decode/encode for BlitFramebuffer.
OpenGL ES 3.x has always done sRGB decode/encode for BlitFramebuffer.

OpenGL 4.x specifications attempted to unify this behavior, and
ultimately settled on doing decode/encode if GL_FRAMEBUFFER_SRGB is
enabled, which allows the full flexibility.  In the process, 4.1,
4.2, early 4.3, and later 4.3 all had somewhat broken rules.

This test was originally written when the latest specification (early
4.3) contained completely contradictory rules, and so simply enforced
a "copy the data unchanged" policy.  This patch updates it to follow
the rules they finally settled on for GL 4.4 and 4.5.

Thanks to Gu Yang and Miao Qiankun for their help analyzing the
differences between the various specifications.

This fixes 20 subtests with NVIDIA 367.35 on a GTX 980:
- blit {renderbuffer,texture} {linear_to_srgb,srgb_to_linear}
  {downsample, msaa, scaled, single_sampled, upsample} enabled

This will break those test cases for Mesa drivers.  I have patches
to make i965 handle the updated rules correctly.

Signed-off-by: Kenneth Graunke <kenneth@whitecape.org>
Reviewed-by: Nicolai Hähnle <nicolai.haehnle@amd.com>
Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>