framework/test/deqp: generalize mustpass list handling
This patch generalizes the mustpass list functionality in deqp_gles3.py.
While the idea is the same, this implementation is considerably
different than the version in deqp_gles3, since it bypasses calling the
binary to generate a test list altogether, and simply replaces the
iter_deqp_test_cases function by providing output that the make_profile
function can consume.

The end result is that this is simpler (though more LOC are added since
this actually has tests), and can be applied to gles2 and gles31.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>