unittests/schema: Drop additionProperties restrction from options
In the past piglit has added or removed options from that part of the
results without caring too much. The reason for this is that these
values are really meant for developers to manually inspect, and for
resuming. We don't support resuming a run on a different version of
piglit than it started (and that's crazy, we never will), so allowing
these values to change arbitrarily is fine.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>