arb_shading_language_420pack: Added more tests for re-declaration in a single declaration
The ARB_shading_language_420pack spec allows multiple appearances of the
same layout qualifier in a single declaration, the last one overriding
any one preceding.

For geometry shaders, the "max_vertices" layout qualifier value has to
be the same in any declaration in the program.

Since the only valid declaration when multiple appearances of the same
layout qualifier is the last one, that wouldn't count as 2 different
declarations. Hence, even if "max_vertices" would have 2 different
values in the same single declaration, that should still be valid.