Add glsl-array-bounds-13.shader_test tests
Compared to the other array bounds tests, this one exercises a larger array
size.

The motivation is for Radeon, where arrays can be lowered either into the
register file or into "real" memory. The larger array in this test will be
sure to hit real memory.

Reviewed-by: Tapani Pälli <tapani.palli@intel.com>
v2: disable the probe (Eric Anholt)