framework: Add a driver classifier based on the renderer string.
For the apitrace test runner, I need to be able to store known-good
images for various drivers, which means I need some way to group
drivers.  Not all Mesa drivers we care about are covered here, but
this is a sample to get things started.

v2: Apply feedback from Dylan, add classification for my Skylake.

Reviewed-by: Dylan Baker <dylan@pnwbakers.com>